+++ src/scene/disk.cpp
@@ -1,23 +1,23 @@
 #include "disk.h"
 
-Disk::Disk(float radius, const Vec3 &position, const Vec3 &normal)
+Disk::Disk(double radius, const Vec3 &position, const Vec3 &normal)
 :n(normal), r(radius), rSquared(r*r)
 {
 	pos = position;
 	d = n.dot(pos);
 }
 
-bool Disk::hit(const Ray &ray, float &t) const
+bool Disk::hit(const Ray &ray, double &t) const
 {
 	// Code is extracted from Plane::hit(...) to avoid too much vfunc calls.
 
-	float dist = (d - ray.origin.dot(n)) / (ray.dir.dot(n));
+	double dist = (d - ray.origin.dot(n)) / (ray.dir.dot(n));
 
 	if(dist > 0 && dist < t)
 	{
 		Vec3 p = ray.origin + ray.dir * dist;
 		p -= pos;
-		float pSquaredL = p.squaredLength();
+		double pSquaredL = p.squaredLength();
 		if (pSquaredL < rSquared)
 		{
 			t = dist;
